pyproject.toml
[project]
name = "manga-translator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.68.0",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
]
